feat: #50 분리했던 StudyHeader 다시 병합
